Improved the documentation
