+++ orbit-2-core/src/main/java/com/babylon/orbit2/syntax/simple/SimpleSyntax.kt
@@ -16,8 +16,6 @@
 
 package com.babylon.orbit2.syntax.simple
 
-import com.babylon.orbit2.Container
-import com.babylon.orbit2.ContainerHost
 import com.babylon.orbit2.syntax.Orbit2Dsl
 import com.babylon.orbit2.syntax.strict.OrbitDslPlugin
 
@@ -29,46 +27,3 @@ class SimpleSyntax<S : Any, SE : Any>(internal val containerContext: OrbitDslPlu
      */
     val state: S get() = containerContext.state
 }
-
-/**
- * Reducers reduce the current state and incoming events to produce a new state.
- *
- * @param reducer the lambda reducing the current state and incoming event to produce a new state
- */
-@Orbit2Dsl
-suspend fun <S : Any, SE : Any> SimpleSyntax<S, SE>.reduce(reducer: suspend SimpleContext<S>.() -> S) {
-    containerContext.apply {
-        reduce { reducerState ->
-            object : SimpleContext<S> {
-                override val state: S = reducerState
-            }.reducer()
-        }
-    }
-}
-
-/**
- * Side effects allow you to deal with things like tracking, navigation etc.
- *
- * There is also a special type of side effects - ones that are meant for the view to listen
- * to as one-off events that are awkward to represent as part of the state - typically things
- * like navigation, showing transient views like toasts etc.
- *
- * These are delivered through [Container.sideEffectFlow] by calling [SimpleSyntax.postSideEffect].
- *
- * @param sideEffect the side effect to post through the side effect flow
- */
-@Orbit2Dsl
-suspend fun <S : Any, SE : Any> SimpleSyntax<S, SE>.postSideEffect(sideEffect: SE) {
-    containerContext.postSideEffect(sideEffect)
-}
-
-/**
- * Build and execute an intent on [Container].
- *
- * @param transformer lambda representing the transformer
- */
-@Orbit2Dsl
-fun <STATE : Any, SIDE_EFFECT : Any> ContainerHost<STATE, SIDE_EFFECT>.intent(transformer: suspend SimpleSyntax<STATE, SIDE_EFFECT>.() -> Unit) =
-    container.orbit {
-        SimpleSyntax(this).transformer()
-    }
